Use 8 bit signed ints in draw circle

+++ draw_circle/draw_circle.cpp
@@ -1,13 +1,14 @@
 #include <api.h>
+#include <inttypes.h>
 
 #include "draw_circle.hpp"
 
 // https://en.wikipedia.org/wiki/Midpoint_circle_algorithm
 
 void draw_circle(unsigned char x_, unsigned char y_, unsigned char r_) {
-  int x = r_;
-  int y = 0;
-  int e = 3 - 2 * r_;
+  int8_t x = r_;
+  int8_t y = 0;
+  int8_t e = 3 - 2 * r_;
   while (x >= y) {
     SET_PIXEL(x_ + x, y_ - y, 1);
     SET_PIXEL(x_ - x, y_ - y, 1);
